Move status and openapi endpoints to actix-web-1
This commit moves the splinterd status, and openapi endpoint functions to
actix-web-1 and removes the serde_derive macro use statement from
splinterd.
